# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/pages/{pageId}/document": {
      "get": {
        "operationId": "getPageDocument",
        "parameters": [
          {
            "name": "pageId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageDocumentDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Page does not exist or belongs to another user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Read the content of a page",
        "tags": ["pages"]
      },
      "put": {
        "operationId": "updatePageDocument",
        "parameters": [
          {
            "name": "pageId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePageDocumentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Document replaced",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageDocumentDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Page does not exist or belongs to another user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Replace the content of a page",
        "tags": ["pages"]
      }
    }
  },
  "info": {
    "title": "Lite Notion API",
    "description": "HTTP API for Lite Notion",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token issued by /auth/login or /auth/register",
        "type": "http"
      }
    },
    "schemas": {
...
      "HttpErrorResponseDto": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Service Unavailable"
          },
          "message": {
            "type": "string",
            "example": "Database is unavailable"
          },
          "path": {
            "type": "string",
            "example": "/api/v1/health"
          },
          "statusCode": {
            "type": "number",
            "example": 503
          },
          "timestamp": {
            "type": "string",
            "example": "2026-08-18T12:00:00.000Z",
            "format": "date-time"
          }
        },
        "required": ["error", "message", "path", "statusCode", "timestamp"]
      },
...
      "PageDocumentDto": {
        "type": "object",
        "properties": {
          "pageId": {
            "type": "string",
            "example": "4c8f1b1a-0f6d-4a5e-9f6d-0f6d4a5e9f6d",
            "format": "uuid"
          },
          "tiptapSchemaVersion": {
            "type": "number",
            "example": 1
          },
          "yjsState": {
            "type": "string",
            "description": "Yjs document state, base64-encoded.",
            "example": "",
            "format": "byte"
          }
        },
        "required": ["pageId", "tiptapSchemaVersion", "yjsState"]
      },
      "UpdatePageDocumentDto": {
        "type": "object",
        "properties": {
          "tiptapSchemaVersion": {
            "type": "number",
            "description": "Tiptap schema version the content conforms to.",
            "example": 1
          },
          "yjsState": {
            "type": "string",
            "description": "Yjs document state, base64-encoded. Opaque to the API. Up to 1048576 bytes.",
            "example": "",
            "format": "byte",
            "maxLength": 1398104
          }
        },
        "required": ["tiptapSchemaVersion", "yjsState"]
      }
    }
  }
}